requirements.txt
Django~=3.2.10
Pillow==8.3.2
django-admin-sortable2==1.0